+++ src/app.ts
@@ -4,7 +4,7 @@ import nodePath from "node:path";
 import { dereference } from "@apidevtools/json-schema-ref-parser";
 import { createHttpTerminator, type HttpTerminator } from "http-terminator";
 
-import { startRepl } from "./repl/repl.js";
+import { startRepl as startReplServer } from "./repl/repl.js";
 import type { Config } from "./server/config.js";
 import { ContextRegistry } from "./server/context-registry.js";
 import { createKoaApp } from "./server/create-koa-app.js";
@@ -20,7 +20,7 @@ import { Transpiler } from "./server/transpiler.js";
 import { CodeGenerator } from "./typescript-generator/code-generator.js";
 
 type MswHandlerMap = {
-  [key: string]: (request: any) => Promise<any>;
+  [key: string]: (request: MockRequest) => Promise<unknown>;
 };
 const allowedMethods = [
   "all",
@@ -61,7 +61,7 @@ export async function createMswHandlers(
 ) {
   // TODO: For some reason the Vitest Custom Commands needed by Vitest Browser mode fail on fs.readFile when they are called from the nested loadOpenApiDocument function.
   // If we "pre-read" the file here it works. This is a workaround to avoid the issue.
-  const _ = await fs.readFile(config.openApiPath);
+  await fs.readFile(config.openApiPath);
   const openApiDocument = await loadOpenApiDocument(config.openApiPath);
   if (openApiDocument === undefined) {
     throw new Error(
@@ -153,13 +153,7 @@ export async function counterfact(config: Config) {
   const koaApp = createKoaApp(registry, middleware, config, contextRegistry);
 
   async function start(options: Config) {
-    const {
-      generate,
-      startRepl: shouldStartRepl,
-      startServer,
-      watch,
-      buildCache,
-    } = options;
+    const { generate, startServer, watch, buildCache } = options;
 
     if (generate.routes || generate.types) {
       await codeGenerator.generate();
@@ -189,11 +183,7 @@ export async function counterfact(config: Config) {
       await transpiler.stopWatching();
     }
 
-    const replServer = shouldStartRepl && startRepl(contextRegistry, config);
-
     return {
-      replServer,
-
       async stop() {
         await codeGenerator.stopWatching();
         await transpiler.stopWatching();
@@ -209,5 +199,6 @@ export async function counterfact(config: Config) {
     koaMiddleware: middleware,
     registry,
     start,
+    startRepl: () => startReplServer(contextRegistry, config),
   };
 }